feat(v7-leverage): 1x/2x/2.5x/3x leverage sweep — 1x MARGINAL, 2x+ FAIL
v5_2 top-3 PLATEAU carrier × 4 leverage × 2 period = 24 backtests.
결론: 1x avg_mdd=-36.4%/cagr=59.7% (MARGINAL), 2x 이상 avg_mdd≥-67% (FAIL).
3x archive 확정. 실거래 = fa80 funding-arb 단독 유지.
runner: --leverage float 지원, SupertrendStrategy: LEVERAGE float 변환.

+++ backtest/strategies/external/SupertrendStrategy.py
@@ -12,7 +12,7 @@ from jesse.strategies import Strategy
 import jesse.indicators as ta
 import os
 
-LEVERAGE = int(os.environ.get('STRATEGY_LEVERAGE', '1'))
+LEVERAGE = float(os.environ.get('STRATEGY_LEVERAGE', '1'))
 
 
 class SupertrendStrategy(Strategy):
